Fix prefetching problem
if unload kernel module *bonding*, l2_bond can not prefetch
data and fails with an error "No such file or directory
- /sys/class/net/bonding_masters".

Closes-bug: #1543609

+++ deployment/puppet/l23network/lib/puppet/provider/l2_base.rb
@@ -500,7 +500,11 @@ class Puppet::Provider::L2_base < Puppet::Provider::InterfaceToolset
     # bond_name => { bond options }
     #
     bond = {}
-    bondlist = File.open("/sys/class/net/bonding_masters").read.chomp.split(/\s+/).sort
+    if File.exist?('/sys/class/net/bonding_masters')
+      bondlist = File.open("/sys/class/net/bonding_masters").read.chomp.split(/\s+/).sort
+    else
+      return {}
+    end
     bondlist.each do |bond_name|
       mode = File.open("/sys/class/net/#{bond_name}/bonding/mode").read.split(/\s+/)[0]
       bond[bond_name] = {
